requirements.txt
absl-py==0.9.0
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
brotlipy==0.7.0
bz2file==0.98
cachetools==4.0.0
certifi==2021.5.30
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
comtypes==1.1.10
cryptography==3.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
entrypoints==0.3
Flask==1.1.1
future==0.18.2
google-auth==1.13.1
google-auth-oauthlib==0.4.1
grpcio==1.28.1
gTTS==2.2.3
gunicorn==20.1.0
h5py==2.10.0
idna==2.10
importlib-metadata==1.5.0
ipykernel==5.1.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
jupyterthemes==0.20.0
kiwisolver==1.2.0
lesscpy==0.14.0
lxml==4.2.5
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
nbconvert==5.6.1
nbformat==5.0.4
nibabel==2.5.1
notebook==6.0.2
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.5.3.56
pandas==1.0.3
pandas-datareader==0.9.0
pandocfilters==1.4.2
parso==0.6.2
pickleshare==0.7.5
Pillow==6.2.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.11.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydicom==1.4.2
Pygments==2.6.1
pyOpenSSL==19.1.0
pyparsing==2.4.6
pypiwin32==223
pyreadline==2.1
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
pyttsx3==2.90
pytz==2019.3
pywin32==227
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.21.3
scipy==1.4.1
Send2Trash==1.5.0
six==1.14.0
soupsieve==2.0.1
tb-nightly==2.3.0a20200406
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.6.0.post2
terminado==0.8.3
testpath==0.4.4
torch==1.2.0
torch-tb-profiler==0.2.1
torchvision==0.3.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.10
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
win-inet-pton==1.1.0
wincertstore==0.2
xlrd==1.2.0
zipp==2.2.0
